plaud_sync_service/requirements.txt
# Azure Functions
azure-functions==1.18.0

# Shared QuickScribe library (installed via editable install)
# Install with: pip install -e ../shared_quickscribe_py

# Azure SDK components (may already be in shared library, but listed for clarity)
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0
azure-storage-queue>=12.8.0
azure-identity>=1.15.0

# Audio processing
ffmpeg-python>=0.2.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Azure Speech Services (swagger client)
# Will be installed from backend/azure_speech/python-client if needed

# Configuration and data formats
PyYAML>=6.0

# Logging and monitoring
opencensus-ext-azure>=1.1.9
